requirements.txt
fastapi
sqlalchemy
alembic
uvicorn
python-dotenv
passlib[bcrypt]
python-jose
fastapi-cors
python-multipart